1_Programing_Language/Python/Practice/Prac_0To120/Practice_109_IMGASST_REMAKE/imageasst/requirements.txt
msgspec==0.18.5
numpy==1.15.1
numpy==1.21.5
Pillow==9.0.1
Pillow==10.2.0
python_magic==0.4.27
pywin32==302
